Refactor to remove all events from reducers. Also, extract all API calls into separate files and remove dependencies
